DropdownColorModeSwitcher component
